requirements.txt
beautifulsoup4
feedparser
gtts
moviepy==1.0.3
nltk
python-dotenv
requests
streamlit